Add a check if a file exists before attempting to read it
Part of the implementation of yarn's new PnP feature is that project dependencies are stored in .zip archives. When coc-nvim tries to read one of those files (for example, during a "goto definition" jump), the `createReadStream` function throws an exception which currently causes Vim to hang completely.

With this change, coc-nvim checks if the file exists before attempting to read it, preventing the thrown exception and preventing Vim from hanging when jumping to a definition located in a PnP module.