feat(diagnostics): show diagnostic message first, hide severity
This is aimed at improving the clarity/readability of the popups.

There's currently a lot of "tags" displayed before the diagnostic
message in the floating window.

The message is usually the most relevant information, but starts partway
through the line and is often wrapped or truncated. Having to hunt for
this info conflicts with the immediate/contextual nature of the popup.

The severity is almost always redundant: the text is colored and the
gutter signs show the severity too.

Before:
```
[clang ovl_no_viable_function_in_call] [E] No matching function for call to 'foo'
```
After:
```
No matching function for call to 'foo'
[clang: ovl_no_viable_function_in_call]
```

(Personally I'd probably prefer to drop the source information from the
hover version entirely, and keep it for `diagnosticPreview` where
there's more space. But that's a more invasive change!)