fix(float): stop floats from getting too small on the RHS of the screen. (#1603)
Currently if a float for diagnostics always triggers to the right
of the cursor. This is fine when the cursor is near column 0, but the right edge
of the window limits the size, so if the cursor is at say 70/80, there
are only 10 columns available and the content wraps *a lot*.

Compounding this, maxheight is set to number of lines of content, but
isn't increased if the content has to wrap. So when the content
wraps, it's truncated. (Scrollable, but this is awkward)

Looking at the history, it seems there's been some tension between
wrapping and fixing widths. 77cbc92ca5c38e1be moved the window to the
left if there wasn't enough room (so it wouldn't wrap), but then
46e0f4d8e5d90ed42 made it conditional on offsetX != 0, which meant
everything except signature help wraps again.

So the proposal here is:
 - on the LHS things stay as they are
 - as the cursor moves to the right, the popup eventually hits the right
   edge, it starts to shrink+wrap
 - once it reaches a minumum width (50 cols, could be a setting), it
   stops shrinking and is no longer locked to the cursor position,
   rather to the RHS of the window.
 - the popup is allowed to grow vertically to accommodate the content
   when it wraps, up to the global maxheight.

Graphically, where the cursor is `X`:

```
  X[------------------]          |
       X[------------------]     |
            X[------------------]|
                 X[-------------]|
                  [----X--------]|
                  [--------X----]|
                  [------------X]|
```