add suggest.maxPreviewHeight
This is a configuration to limit the max height of floating windows AND
preview windows. This is especially useful when hoverTarget is preview
and previewAutoClose is false.  The Main advantage of such configuration
is that the documentation text remains intact while editing/navigating.
However, if the documentation is very long, the preview window would
occupy the entire page:
```
// in provideTextDocumentContent
nvim.command(`exe "normal! z${this.documentLines.length}\\<cr>"`, true)
```
Users can prevent such cases by setting maxPreviewHeight to a value
small enough (e.g. 20).

The name of option doesn't look good because there is another option
with the exactly same name (list.maxPreviewHeight).  Something like
suggest.hoverMaxHeight and suggest.hoverMaxWidth would be better.

TODO:
It turns out that this actually doesn't limit the max height of floating
window and I couldn't figure out why.