zN<CR> only if preview window is a horizontal split (#2306)
When using "coc.preferences.hoverTarget": "preview", coc sets the
preview window height with zN<CR>. The problem is that it reduces the
height of the window even if there is no window adjacent to the
top/bottom current window, which is undesirable. Therefore, it first
need to check that there is a adjacent window.