Add link to my detailed article on coc extensions
I wrote this article because I struggled a bit writing my first coc
extension to wrap an executable language server. The article provides
both the reason and the method to write a coc.nvim extension.
Additionally, it goes over why coc.nvim is different from other nvim LSP
clients and how it relates to vscode. My hope is that it makes it easier
for other developers to write coc extensions so that we get more of
them.