popup: escape markdown syntax
this commit removes the escapes from markdown returned from a lsp on
hover.

the timing of CocFloatOpen will not work. Instead I tie an autocmd to
the action and call it when CocAction completes.

this pr opens up the use of registering autocmd(s) named after a
CoCAction and having those autocmd fire when the rpc request actually
returns to (n)vim.