feat: use sign group
Coc tracked which signs it had placed to clear those later, but this
list is frequently out of sync with what actually shown on screen.

By placing all of Coc sign in their own group, we could clear all of
those signs with 1 function call.

checkSigns method is used as workaround for this problem thus is no
longer needed.