fix(highlight): revert winhighlight purge for some floats
These specific floats (button, border, etc.) shouldn't carry syntax
highlighting, so purging them from winhighlight most probably shouldn't
bring practical benefits. At the same time the window local highlight
fix has a side effect (see PR #3487) which is only problematic for
windows that get focus most of the time, which is the case for these
kinds of windows (buttons, etc.).

So we keep usage of winhighlight for these floats in specific (even
thought ideally it would be better to have a full purge) due to the side
effect of the current solution.