Feat(rename.ts): Tweak To Refactor-Rename Floating Window (#3528)
* feat(rename.ts): Configurable Refactor/Rename Floating Window

When renaming, I find it more "natural" for typing to have the user
input window display without the previous name. This removes the need to
hold delete to remove the old name, before entering a new one.
I've add a small conditional statement to check for a new config key
`coc.preferences.fillRename`, to decide whether to pass a default value
to `window.requestInput()`. Default is true, which keeps the
functionality the same as what we currently have. False means we don't
pass the old name as a default value to requestInput window, so the
floating window is empty and the user can immediately enter a new name.

* docs(coc.preferences): Added fillRename documentation

* feat(rename.ts): Changed config key to renameFillCurrent