:CheckHealth should handle Vim
`:CheckHealth` is most popular in the `nvim` world, but can also be used
on the vim side when rhysd's excellent `vim-healthcheck` plugin is
installed: https://github.com/rhysd/vim-healthcheck.git

This is a minor change, one that allows the health checks to properly
handle vim and determine if it is new enough.  For the minimum version
checks I used the current values examined in `plugins/coc.vim`.

(This change doesn't address if the checks should report on the
differences between the minimum supported and the optimal minimum.)