feat(list): expose listItems for extensions
Every extension can register a source for list, but other extensions
must call `CocAction('listLoadItems', name)` to get the list items by
name. However, some sources like `symbols` may return a ton of data.
For a extension to get data, Neo/Vim and nodeJS must encode & decode
for the data which is very inefficient.

Solution:
We can expose a internal command for the extension, which is
efficient and more friendly to pure Typescript.