fix(files): Workaround Windows path limitation on Vim too (#3913)
This improves the workaround introduced in #3904 and #3907. The previous
implementation is using Neovim-only API, but new implementation uses
`bufadd()` to create the buffer with specified name, so it works on Vim
too. It also simplifies code too.

With quick testing, it seems that new implementation works without issue
like #3906 even if non-URI like path was given, so you may not have to
check it looks like URI, but keep it for now to reduce the risk of
potential breakage.