make tagfunc cmds silent (#4402)
previously, the commands generated by the CoC tagfunc produced output that caused the press-enter dialog to appear when selecting from multiple options with g<C-]>. make the commands silent so that this doesn't happen.

the commands themselves are still displayed in the g<C-]> dialog when there are multiple options, which may potentially confuse a user. a better choice might be to replace cmd with the exact string that matches the tag, like the native tagfunc does.