fix(codeLens): Misalign when using tabs as indent
From vim doc [1]:
Any Tab and other control character in the text will be changed to a
space (Rationale: otherwise the size of the text is difficult to
compute).

If we use tabs as indent, then the code lens will be misaligned with the
real text even if we set the indent of the code lens to the right number
of tabs.

[1] https://vimhelp.org/textprop.txt.html#virtual-text

Signed-off-by: Adam Tao <tcx4c70@gmail.com>